+++ src/main/java/shop/nuribooks/books/controller/category/CategoryController.java
@@ -12,8 +12,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
-import shop.nuribooks.books.dto.category.request.CategoryRequest;
-import shop.nuribooks.books.dto.category.response.CategoryResponse;
+import shop.nuribooks.books.dto.category.CategoryRequest;
+import shop.nuribooks.books.dto.category.CategoryResponse;
 import shop.nuribooks.books.service.category.CategoryService;
 
 /**
